pyproject.toml
[project]
name = "peakrdl-rust"
version = "0.5.0"
description = "Generate a Rust crate from SystemRDL for accessing control/status registers."
readme = "README.md"
authors = [{ name = "Dana Sorensen", email = "dana.r.sorensen@gmail.com" }]
requires-python = ">=3.9"
license = "LGPL-2.1-or-later"
keywords = [
    "SystemRDL",
    "PeakRDL",
    "CSR",
    "compiler",
    "tool",
    "registers",
    "generator",
    "Rust",
    "PAC",
    "embedded",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "License :: OSI Approved :: GNU Lesser General Public License v2 (LGPLv2)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Rust",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Typing :: Typed",
]
dependencies = [
    "case-converter>=1.2.0",
    "jinja2>=3.1.6",
    "systemrdl-compiler>=1.29.3",
]

[project.urls]
Documentation = "https://peakrdl-rust.readthedocs.io/"
Source = "https://github.com/darsor/PeakRDL-rust"
Tracker = "https://github.com/darsor/PeakRDL-rust/issues"
Changelog = "https://github.com/darsor/PeakRDL-rust/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "mypy>=1.8.0",
    "peakrdl>=1.4.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "pytest>=8.4.2",
    "ruff>=0.3.0",
]
docs = [
    "pygments-systemrdl>=1.3.0",
    "sphinx>=4.0.0",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-book-theme>=0.3.0",
    "sphinxcontrib-spelling>=8.0.0",
]

[project.optional-dependencies]
cli = [
    "peakrdl-cli>=1.4.0"
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.pyright]
typeCheckingMode = "standard"
include = ["src"]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# ignore missing type information for caseconverter dependency
[[tool.mypy.overrides]]
module = ["caseconverter.*"]
ignore_missing_imports = true

[project.entry-points."peakrdl.exporters"]
rust = "peakrdl_rust.__peakrdl__:Exporter"
